Refactor directory to accomodate extensions Lovelace is now a package! It requires a different command to run, the README is updated. There is now an overaching bot directory folder with appropriate __init__ and __main__ files. Additionally, there are two new folder strucutres: exts and utils. exts is for Extensions while util is for utility functions and classes. Functionality for autoloading extensions was also added. It checks if a setup function is in the file, if yes than it'll load in the extension.
